refactor: migrate add to calendar functionality to use callback registry; remove deprecated config-add-to-calendar file
