+++ packages/app-sundevil/src/components/GameTableSection/add-to-calendar/add-to-calendar-callback-registry.js
@@ -1,0 +1,18 @@
+export class AddToCalendarCallbackRegistry {
+  constructor() {
+    this.callback = null;
+  }
+
+  async call() {
+    if (typeof this.callback === "function") {
+      return this.callback?.();
+    }
+    return null;
+  }
+
+  async register(callback) {
+    this.callback = callback;
+  }
+}
+
+window.AddToCalendarCallbackRegistry = AddToCalendarCallbackRegistry;
